requirements.txt
autopep8==1.6.0
azure-common==1.1.27
azure-core==1.21.1
azure-servicebus==7.5.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
click==8.0.3
cryptography==3.4.7
docopt==0.6.2
flake8==4.0.1
Flask==2.0.0
Flask-BasicAuth==0.2.0
gunicorn==20.1.0
idna==2.10
importlib-metadata==4.2.0
isodate==0.6.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mccabe==0.6.1
msrest==0.6.21
oauthlib==3.1.1
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
requests==2.24.0
requests-oauthlib==1.3.0
six==1.16.0
toml==0.10.2
typing-extensions==4.0.1
uamqp==1.5.1
urllib3==1.25.11
Werkzeug==2.0.2
zipp==3.7.0
